Split tests for page where functionality
