+++ saleor/graphql/page/tests/queries/pages_with_where/shared.py
@@ -1,0 +1,12 @@
+QUERY_PAGES_WITH_WHERE = """
+    query ($where: PageWhereInput) {
+        pages(first: 5, where:$where) {
+            totalCount
+            edges {
+                node {
+                    id
+                }
+            }
+        }
+    }
+"""
